Added a NewVersion of the GUI, New Tab for View in non table format.
Need to add a new query for this table as it currently uses the custom
one. Need to add a feature to add new item and also to remove etc.

I have tried to add a 3D model viewer but this is not working yet. May
leave this for now and add it for my next time.

Also need to integrate into Maya and Houdini.